dcp-t5(004): move cli/initsetup under operator_settings
Relocate pkg/transports/cli/initsetup to
pkg/services/operator_settings/transports/cli/initsetup. Its production
code imports only operator_settings among product services, so it belongs
in that service's own transports subtree.

Pure relocation: contents are unchanged beyond import paths in four
external importers plus the self-import in init_test.go. The package's
pre-existing test-only dependency on pkg/services/providers/wire is
deliberately preserved: it is confined to init_test.go, and both
cmd/servicecyclecheck (graph.go) and the peer-import walk in
cmd/ownershipboundarycheck skip _test.go files, so it contributes no
cross-service edge and no boundary finding. Unit and functional coverage
minimums move with the package.
