+++ pkg/services/operator_settings/transports/cli/initsetup/init.go
