+++ pkg/services/operator_settings/transports/cli/initsetup/testmain_test.go
